+++ include/ryser_algorithms.h
@@ -3,11 +3,11 @@
 #include "matrix_utils.h"
 long long computePermanentRyserGreyCodeSparse(const std::vector<NonZeroElement>& nonZeroElements, int n);
 
-long long computePermanentRyserGreyCode(int* A, int n);
+long long computePermanentRyserGreyCode(double* A, int n);
 
-long long computePermanentRyser(int* A, int n);
+long long computePermanentRyser(double* A, int n);
 
-long long computePermanentRyserPar(int* A, int n);
+long long computePermanentRyserPar(double* A, int n);
 
 long long computePermanentRyserSparsePar(const std::vector<NonZeroElement>& nonZeroElements, int n);
